extract prompts from CT generator
