+++ packages/generators/generators/lib/plops/utils/get-draft-and-publish-prompts.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = [
+  {
+    type: 'confirm',
+    name: 'useDraftAndPublish',
+    default: false,
+    message: 'Use draft and publish?',
+  },
+];
