+++ packages/generators/generators/lib/plops/utils/get-kind-prompts.js
@@ -1,0 +1,17 @@
+'use strict';
+
+const validateInput = require('./validate-input');
+
+module.exports = [
+  {
+    type: 'list',
+    name: 'kind',
+    message: 'Please choose the model type',
+    default: 'collectionType',
+    choices: [
+      { name: 'Collection Type', value: 'collectionType' },
+      { name: 'Single Type', value: 'singleType' },
+    ],
+    validate: input => validateInput(input),
+  },
+];
